Redact Sensitive Data From Request Payload Logs
