+++ internal/broker/logging.go
@@ -1,7 +1,10 @@
 package broker
 
 import (
+	"context"
 	"encoding/json"
+	"fmt"
+	"log/slog"
 	"reflect"
 )
 
@@ -10,7 +13,6 @@ var openKeys = map[string]struct{}{
 	"xsappname":        {},
 	"globalaccount_id": {},
 	"subaccount_id":    {},
-	"kubeconfig":       {},
 }
 
 func hideSensitiveDataFromRawContext(d []byte) map[string]interface{} {
@@ -60,4 +62,63 @@ func marshallRawContext(d map[string]interface{}) string {
 		return "unable to marshal context data"
 	}
 	return string(b)
+}
+
+// strippingHandler is a slog.Handler that truncates attributes with the given keys
+// to maxInstanceDetailsLen characters. Prevents brokerapi's "instance-details"
+// attribute (which embeds the full raw request payload) from flooding test output.
+const maxInstanceDetailsLen = 1024
+
+type strippingHandler struct {
+	inner slog.Handler
+	strip map[string]struct{}
+}
+
+func NewStrippingHandler(inner slog.Handler, keys ...string) *strippingHandler {
+	m := make(map[string]struct{}, len(keys))
+	for _, k := range keys {
+		m[k] = struct{}{}
+	}
+	return &strippingHandler{inner: inner, strip: m}
+}
+
+func (h *strippingHandler) Enabled(ctx context.Context, level slog.Level) bool {
+	return h.inner.Enabled(ctx, level)
+}
+
+func (h *strippingHandler) Handle(ctx context.Context, r slog.Record) error {
+	filtered := slog.NewRecord(r.Time, r.Level, r.Message, r.PC)
+	r.Attrs(func(a slog.Attr) bool {
+		if _, truncate := h.strip[a.Key]; truncate {
+			s := fmt.Sprintf("%v", a.Value.Any())
+			if len(s) > maxInstanceDetailsLen {
+				s = s[:maxInstanceDetailsLen] + "...[truncated]"
+			}
+			filtered.AddAttrs(slog.String(a.Key, s))
+		} else {
+			filtered.AddAttrs(a)
+		}
+		return true
+	})
+	return h.inner.Handle(ctx, filtered)
+}
+
+func (h *strippingHandler) WithAttrs(attrs []slog.Attr) slog.Handler {
+	filtered := make([]slog.Attr, 0, len(attrs))
+	for _, a := range attrs {
+		if _, skip := h.strip[a.Key]; !skip {
+			filtered = append(filtered, a)
+		} else {
+			s := fmt.Sprintf("%v", a.Value.Any())
+			if len(s) > maxInstanceDetailsLen {
+				s = s[:maxInstanceDetailsLen] + "...[truncated]"
+			}
+			filtered = append(filtered, slog.String(a.Key, s))
+		}
+	}
+	return &strippingHandler{inner: h.inner.WithAttrs(filtered), strip: h.strip}
+}
+
+func (h *strippingHandler) WithGroup(name string) slog.Handler {
+	return &strippingHandler{inner: h.inner.WithGroup(name), strip: h.strip}
 }
